Set Docker build to Release, add DatingApp.Common.csproj
The Dockerfile now explicitly sets the build configuration to `Release`, ensuring that the application is built in release mode.

The `DatingApp.Common.csproj` file is now included in the Docker image, indicating it is a new dependency for the application.

Removed a commented-out line that previously copied the `DatingApp.Api.csproj` file, improving readability and maintainability of the Dockerfile.