Remove outdated URLs from app.Urls in development mode
The URLs "https://dattingapp-datingapp.piemei.easypanel.host" and
"http://dattingapp-datingapp.piemei.easypanel.host" have been removed
from the `app.Urls` collection in `Program.cs` within the
`DatingApp.Api` namespace. This change occurs within a conditional
block that checks if the environment is in development mode. The
removal is likely part of a configuration update to eliminate outdated
or incorrect URLs.